# Don't store graph offsets for HNSW graph