… lucene/core/src/test/org/apache/lucene/util/hnsw/KnnGraphTester.java