@msokolov 2021-12-13
I seem to remember that when I checked (you can use `-fanout` parameter to `KnnGraphTester` IIRC) most nodes were *not* fully populated; ie they had fewer than `maxConn` connections. Why? This seems counterintuitive, but I think it would be prudent to check the size increase/decrease from this change for some dataset/parameter choices. Certainly it's appealing to avoid the extra offset/lookup data structure