@mayya-sharipova 2021-12-14
@msokolov Thanks for the initial review, it is good to know that we are ok with this idea. I will do the comparison of size and also the `maxConn` numbers.