@mayya-sharipova 2022-01-05
I've also run the comparison on a bigger dataset: deep-image-96-angular of 10M docs.
M: 16; efConstruction: 500

Disk size before the change: 4.2G; after change: 4.3G => 2% increase
Not much affect on search performance:

|             | baseline recall | baseline QPS | candidate recall | candidate QPS |
| ----------- | --------------: | -----------: | ---------------: | ------------: |
| n_cands=10  |           0.726 |     1527.894 |            0.728 |       870.721 |
| n_cands=20  |           0.793 |     1350.206 |            0.794 |      1364.301 |
| n_cands=40  |           0.862 |     1053.906 |            0.862 |      1068.798 |
| n_cands=80  |           0.917 |      737.711 |            0.918 |       741.551 |
| n_cands=120 |           0.942 |      573.783 |            0.942 |       589.756 |
| n_cands=200 |           0.964 |      402.166 |            0.964 |       414.730 |
| n_cands=400 |           0.982 |      237.545 |            0.982 |       251.678 |
| n_cands=600 |           0.988 |      174.223 |            0.988 |       177.968 |
| n_cands=800 |           0.991 |      137.420 |            0.991 |       143.290 |