@jtibshirani 2022-01-10
Oh sorry, I didn't see this was merged into the `hnsw` branch. Ignore my comment, I guess we'll handle the new format in another PR.