@mayya-sharipova 2022-01-11
@jtibshirani Thanks for the guide on the format change, I will study it and follow it.
Indeed this PR was merged into the `hnsw` branch, so we will do the format change there.